fixs

+++ database/migrations/2021_07_08_095525_create_benefits_proposals_table.php
@@ -19,7 +19,7 @@ class CreateBenefitsProposalsTable extends Migration
             $table->integer('benefit_id')->unsigned();
             $table->integer('proposal_id')->unsigned();
             $table->string('name')->nullable();
-            $table->integer('value')->nullable();
+            $table->float('value')->nullable();
             $table->string('type')->nullable();
             $table->timestamps();
             $table->softDeletes();
